Reorganised the file structure.
